Move stuff into bin and lib dirs
