+++ protobuf_benchmarks/compile_protos.sh
@@ -26,10 +26,9 @@ PROTOS=(
 )
 
 SCRIPT_DIR=$(dirname "${BASH_SOURCE}")
-OUTPUT_DIR="${SCRIPT_DIR}/temp"
-
-mkdir -p ${OUTPUT_DIR}
 
 set -x
 
-protoc -I"${SCRIPT_DIR}" --dart_out="${OUTPUT_DIR}" --plugin=protoc-gen-dart=run_protoc_plugin.sh "${PROTOS[@]/#/$SCRIPT_DIR/}"
+mkdir -p lib/generated
+
+protoc -I"${SCRIPT_DIR}" --dart_out=lib/generated --plugin=protoc-gen-dart=run_protoc_plugin.sh "${PROTOS[@]/#/$SCRIPT_DIR/}"
